First version 0.50.0, with breaking change
This version has a breaking change in the library. The NetworkElement
class holds the type of the state to avoid unchecked casts, change was
in commit: 38ab05ffd6f968377b229cebf28655b1f186746d

As I did not keep track of versioning before, I will name this 0.50.0
using semantic versioning 2.0.0. As this is a prerelease, the minor
version might might break API functionality, and patch version should
not.