Add Artifactory configuration for deployment
Artifactory is a local repository where all the other projects will
depend on to manage the dependencies. It will let me keep track of
the different versions of my libraries and whether they break changes.

Additionally, jars with the compiled classes and source are provided to
use as a standalone library without version tracking.